readme updates

+++ install/install-megacc.sh
@@ -8,7 +8,7 @@ if [ $ENV == "native" ]; then
     wget https://www.megasoftware.net/do_force_download/megax-cc_10.2.6-1_amd64.deb
     apt-get install -y ./megax-cc_10.2.6-1_amd64.deb
     rm megax-cc_10.2.6-1_amd64.deb
-else
+elif [ $ENV == "docker" ]; then
     wget https://megasoftware.net/do_force_download/mega-cc_11.0.13-1_amd64.deb
     apt-get install -y ./mega-cc_11.0.13-1_amd64.deb
     rm mega-cc_11.0.13-1_amd64.deb
